refactor(uuid): factor out common symbols into `common` module
