+++ uuid/common.ts
@@ -1,0 +1,56 @@
+// Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
+// This module is browser compatible.
+
+import { NIL_UUID } from "./constants.ts";
+
+/**
+ * Check if the passed UUID is the nil UUID.
+ *
+ * @example
+ * ```ts
+ * import { isNil } from "@std/uuid";
+ *
+ * isNil("00000000-0000-0000-0000-000000000000"); // true
+ * isNil(crypto.randomUUID()); // false
+ * ```
+ */
+export function isNil(id: string): boolean {
+  return id === NIL_UUID;
+}
+
+/**
+ * Test a string to see if it is a valid UUID.
+ *
+ * @example
+ * ```ts
+ * import { validate } from "@std/uuid";
+ *
+ * validate("not a UUID"); // false
+ * validate("6ec0bd7f-11c0-43da-975e-2a8ad9ebae0b"); // true
+ * ```
+ */
+export function validate(uuid: string): boolean {
+  return /^(?:[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}|00000000-0000-0000-0000-000000000000)$/i
+    .test(
+      uuid,
+    );
+}
+
+/**
+ * Detect RFC version of a UUID.
+ *
+ * @example
+ * ```ts
+ * import { version } from "@std/uuid";
+ *
+ * version("d9428888-122b-11e1-b85c-61cd3cbb3210"); // 1
+ * version("109156be-c4fb-41ea-b1b4-efe1671c5836"); // 4
+ * ```
+ */
+export function version(uuid: string): number {
+  if (!validate(uuid)) {
+    throw new TypeError("Invalid UUID");
+  }
+
+  return parseInt(uuid[14]!, 16);
+}
