+++ uuid/constants.ts
@@ -45,3 +45,8 @@ export const NAMESPACE_OID = "6ba7b812-9dad-11d1-80b4-00c04fd430c8";
  * ```
  */
 export const NAMESPACE_X500 = "6ba7b814-9dad-11d1-80b4-00c04fd430c8";
+/**
+ * The nil UUID is special form of UUID that is specified to have all 128 bits
+ * set to zero.
+ */
+export const NIL_UUID = "00000000-0000-0000-0000-000000000000";
