+++ uuid/deno.json
@@ -3,6 +3,7 @@
   "version": "0.224.0",
   "exports": {
     ".": "./mod.ts",
+    "./common": "./common.ts",
     "./constants": "./constants.ts",
     "./v1": "./v1.ts",
     "./v3": "./v3.ts",
